Vuelve la fecha de validacion a null cuando se edita un perfume

+++ src/main/webapp/views/frontoffice/perfumes.jsp
@@ -39,7 +39,7 @@
 					<td>${perfume.precio}</td>
 					<td><img class="img-thumbnail img-table" alt="imagen perfume" src="${perfume.imagen}"></td>
 					<td class="blockquote text-center"><a href="views/frontoffice/crear_perfume?id=${perfume.id}"> <i class="fas fa-edit fa-2x"></i></a>
-						<a href="views/frontoffice/eliminar_perfume?id=${perfume.id}"><i class="fas fa-trash fa-2x"></i></a></td>
+						<a href="views/frontoffice/eliminar_perfume?id=${perfume.id}" onclick="confirmar('${perfume.nombre}')"><i class="fas fa-trash fa-2x"></i></a></td>
 				</tr>
 			</c:forEach>
 		</tbody>
